Renamed color map => theme
