+++ lib/Expression/Theme/Util/Gradient.php
@@ -1,9 +1,10 @@
 <?php
 
-namespace PhpBench\Expression\ColorMap\Util;
+namespace PhpBench\Expression\Theme\Util;
 
 use function array_fill;
 use RuntimeException;
+use PhpBench\Expression\Theme\Util\Color;
 
 final class Gradient
 {
